+++ src/mol-canvas3d/passes/image.ts
@@ -35,6 +35,7 @@ export const ImageParams = {
 
     cameraHelper: PD.Group(CameraHelperParams),
     renderer: PD.Group(RendererParams),
+    transparency: PD.Select('blended', PD.arrayToOptions(['blended', 'wboit', 'dpoit'] as const))
 };
 export type ImageProps = PD.Values<typeof ImageParams>
 
@@ -57,10 +58,10 @@ export class ImagePass {
     get width() { return this._width; }
     get height() { return this._height; }
 
-    constructor(private webgl: WebGLContext, assetManager: AssetManager, private renderer: Renderer, private scene: Scene, private camera: Camera, helper: Helper, transparency: 'wboit' | 'dpoit' | 'blended', props: Partial<ImageProps>) {
+    constructor(private webgl: WebGLContext, assetManager: AssetManager, private renderer: Renderer, private scene: Scene, private camera: Camera, helper: Helper, props: Partial<ImageProps>) {
         this.props = { ...PD.getDefaultValues(ImageParams), ...props };
 
-        this.drawPass = new DrawPass(webgl, assetManager, 128, 128, transparency);
+        this.drawPass = new DrawPass(webgl, assetManager, 128, 128, this.props.transparency);
         this.illuminationPass = new IlluminationPass(webgl, this.drawPass);
         this.multiSamplePass = new MultiSamplePass(webgl, this.drawPass);
         this.multiSampleHelper = new MultiSampleHelper(this.multiSamplePass);
@@ -100,6 +101,7 @@ export class ImagePass {
 
     setProps(props: Partial<ImageProps> = {}) {
         Object.assign(this.props, props);
+        if (props.transparency) this.drawPass.setTransparency(props.transparency);
         if (props.cameraHelper) this.helper.camera.setProps(props.cameraHelper);
     }